Section 5: 119-124
Created Tank.cpp and Tank.h based on BasePawn
BasePawn.h: Updated UPROPERTY options of private variables. REmoved SetupPlayerInputComponent().
BPTower and BPTank were updated with respective meshes.
Added into the level Tank and Towers
Tank.h: Added construtor, SpringArm and Camera. Added SetupPlayerInputComponent() and Move()
Tank.cpp: Initialized constructor, SpringArmm attached to the rootComponent and camera to the SpringArm. In Move() we print the values of inputs in the log. SetupPlayerInputComponent gives the values from the inputs "MoveForward" to Move().
Assigned the tank on the map the Character 0