Section 5: 139 - 141
Added in Projectile.cpp appplyDamage()
In BasePawn.cpp assigned the projectile as his own Owner
In HealthComponent reduced he health every time there is damage
Created GameModeBase class and blueprint called ToonTanksGameMode
Assigned the BP_ToonTanksGameMode as default gamemode in project settings
Handled Actor death. Added function to handle death on tank and tower, as well in gamemode. In healhcomponent, manage the dead of the actor according to what actor it is.